+++ Interfaces/ITokenGenerator.cs
@@ -1,0 +1,7 @@
+﻿namespace AuthService.Interfaces
+{
+    public interface ITokenGenerator
+    {
+        public string GenerateRefreshToken();
+    }
+}