+++ Security/TokenGenerator.cs
@@ -1,8 +1,9 @@
 ﻿using System.Security.Cryptography;
+using AuthService.Interfaces;
 
 namespace AuthService.Security
 {
-    public class TokenGenerator
+    public class TokenGenerator : ITokenGenerator
     {
         public string GenerateRefreshToken()
         {