+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ﻿# CHANGELOG
 
+## AuthService V 1.0.2 [29/03/26]
+### Added
+- Added interfaces for the services
+### FIXED	
+- AuthService.Service.AuthService renamed to AuthService.Service.AuthManager
+
+
 ## AuthService V 1.0.1 [29/03/26]
 ### Added
 - Added CHANGELOG